changed input handling with scene2d, created controlActor and OptionActor for this handling, created the level selection scene

+++ v0_2/core/src/com/plearning/game/PlearningGame.java
@@ -20,8 +20,8 @@ public class PlearningGame extends Game {
 	
 	FPSLogger fpsLogger; //prints fps in-game, only in dev
 	
-	float scaleFactorX;
-	float scaleFactorY;
+	int scaleFactorX;
+	int scaleFactorY;
 	
 	AssetManager manager;
 	public boolean soundEnabled;
